+++ src/Blobject.CIFS/CifsBlobClient.cs
@@ -235,6 +235,12 @@ namespace Blobject.CIFS
         }
 
         /// <inheritdoc />
+        public override async Task<DeleteManyResult> DeleteManyAsync(IEnumerable<string> keys, CancellationToken token = default)
+        {
+            return await base.DeleteManyAsync(keys, token).ConfigureAwait(false);
+        }
+
+        /// <inheritdoc />
         public override async Task<bool> ExistsAsync(string key, CancellationToken token = default)
         {
             string path = BuildFilePath(key);