+++ src/Blobject.Disk/DiskBlobClient.cs
@@ -236,6 +236,12 @@ namespace Blobject.Disk
         }
 
         /// <inheritdoc />
+        public override async Task<DeleteManyResult> DeleteManyAsync(IEnumerable<string> keys, CancellationToken token = default)
+        {
+            return await base.DeleteManyAsync(keys, token).ConfigureAwait(false);
+        }
+
+        /// <inheritdoc />
         public override async Task<bool> ExistsAsync(string key, CancellationToken token = default)
         {
             string filename = GenerateUrl(key);