+++ src/Blobject.GoogleCloud/GcpBlobClient.cs
@@ -232,6 +232,12 @@
         }
 
         /// <inheritdoc />
+        public override async Task<DeleteManyResult> DeleteManyAsync(IEnumerable<string> keys, CancellationToken token = default)
+        {
+            return await base.DeleteManyAsync(keys, token).ConfigureAwait(false);
+        }
+
+        /// <inheritdoc />
         public override async Task<bool> ExistsAsync(string key, CancellationToken token = default)
         {
             try